:sparkles: refactor landing page components
