+++ src/pages/landing-page/subpages/smartpanel/components/CardContainer.tsx
@@ -1,0 +1,37 @@
+import { Card, SxProps } from "@mui/material";
+import { FC, ReactNode } from "react";
+import { border, color } from "../../../../../styles/constants";
+
+interface CardContainerProps {
+  children: ReactNode;
+  containerStyle?: SxProps;
+}
+
+const CardContainer: FC<CardContainerProps> = ({
+  children,
+  containerStyle,
+}) => (
+  <Card
+    elevation={0}
+    sx={{
+      height: "100%",
+      width: "100%",
+      display: "flex",
+      flexDirection: "column",
+      justifyContent: "center",
+      alignItems: "center",
+      bgcolor: color.white.twoTenTransparent,
+      border: `${border.xs} ${color.brightBlue.semiTransparentDark}`,
+      color: "#fff",
+      "&:hover": {
+        scale: "101%",
+        cursor: "pointer",
+      },
+      ...containerStyle,
+    }}
+  >
+    {children}
+  </Card>
+);
+
+export default CardContainer;
